+++ packages/frontend/editor-ui/src/features/ai/chatHub/chat.utils.ts
@@ -572,3 +572,25 @@ export function isWaitingForApproval(message: ChatMessage | null | undefined): b
 
 	return message.content.some((c) => c.type === 'with-buttons' && c.blockUserInput);
 }
+
+export function chunkFilesBySize(files: File[], maxSizeBytes: number): File[][] {
+	const chunks: File[][] = [];
+	let currentChunk: File[] = [];
+	let currentSize = 0;
+
+	for (const file of files) {
+		if (currentSize + file.size > maxSizeBytes && currentChunk.length > 0) {
+			chunks.push(currentChunk);
+			currentChunk = [];
+			currentSize = 0;
+		}
+		currentChunk.push(file);
+		currentSize += file.size;
+	}
+
+	if (currentChunk.length > 0) {
+		chunks.push(currentChunk);
+	}
+
+	return chunks;
+}
